Get imgui up and running. Add AutoC tool to play with autogenerating function declarations.

+++ src/platform/glfw/glfw.h
@@ -53,7 +53,7 @@ Link lib
 FUNCTIONS
 =========================================================*/
 
-uint32_t glfw__get_time(platform_t* platform);
+float glfw__get_time(platform_t* platform);
 
 boolean glfw__load_file(const char* filename, boolean binary, long* out__size, void** out__buffer);
